+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bugfixes
 
 - Fix panic if multiple instances are configured and too many instances become unavailable ([#313](https://github.com/Nitrokey/nethsm-pkcs11/issues/313))
+- Temporarily disable parallel key listing on Windows.
 
 ## [2.1.0][] (2026-02-27)