+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+-
+
+## [2.2.0][] (2026-04-23)
+
 ### Features
 
 - Add `disable_threads` and `disable_thread_pool` config options
@@ -13,6 +17,9 @@
   - Remove unnecessary version check so that the argument is handled at all.
   - Correctly handle `null` arguments and initialize the thread state consistently.
   - Return an error if some but not all of the `CreateMutex`, `DestroyMutex`, `LockMutex` and `UnlockMutex` fields are set.
+
+[2.2.0]: https://github.com/Nitrokey/nethsm-pkcs11/releases/tag/v2.2.0
+[Full Changelog](https://github.com/Nitrokey/nethsm-pkcs11/compare/v2.1.0...v2.2.0)
 
 ## [2.1.0][] (2026-02-27)